missed files

+++ lib/bauble/cli/code_bundler.rb
@@ -1,7 +1,9 @@
 # frozen_string_literal: true
 
 require 'logger'
+require 'English'
 require_relative 'docker_command_builder'
+require_relative 'bundle_command_builder'
 
 module Bauble
   module Cli
@@ -14,30 +16,46 @@ module Bauble
               Logger.docker(line)
             end
           end
+
+          return if $CHILD_STATUS.success?
+
           `rm -rf .bundle`
+          Logger.error('Bundle step failed')
+          exit
         end
 
         private
 
         # TODO: Remove the need for this to install things from the sub dir
         def docker_build_gems_command(bundle_hash)
-          DockerCommandBuilder
-            .new
-            .with_rm
-            .with_volume('$(pwd)/../:/var/task')
-            .with_workdir('/var/task/demo_app')
-            .with_entrypoint('/bin/sh')
-            .with_platform('linux/amd64')
-            .with_image('public.ecr.aws/sam/build-ruby3.2')
-            .with_command(bundle_command(bundle_hash))
-            .build
+          command = DockerCommandBuilder
+                    .new
+                    .with_rm
+                    .with_volume('$(pwd):/var/task')
+                    .with_workdir('/var/task')
+                    .with_entrypoint('/bin/sh')
+                    .with_platform('linux/amd64')
+
+          if ENV['BAUBLE_DEV_GEM_PATH']
+            command = command.with_volume("#{ENV['BAUBLE_DEV_GEM_PATH']}:/var/task/bauble_core")
+          end
+
+          command.with_image('public.ecr.aws/sam/build-ruby3.2')
+                 .with_command(bundle_command(bundle_hash))
+                 .build
         end
 
         def bundle_command(bundle_hash)
-          'bundle config set without test development && ' \
-          "bundle config set path \".bauble/assets/#{bundle_hash}/gem-layer\" && " \
-          'bundle install && ' \
-          'rm -rf .bundle'
+          command = BundleCommandBuilder
+                    .new
+                    .with_bundle_without(%w[test development])
+                    .with_bundle_path(bundle_hash)
+
+          command.with_bauble_gem_override if ENV['BAUBLE_DEV_GEM_PATH']
+
+          command.with_bundle_install
+                 .with_dot_bundle_cleanup
+                 .build
         end
       end
     end
